refactor(range-selector): remove direction handling from MobileRangeSelector and RangeSelector
